+++ CHANGELOG.md
@@ -46,6 +46,7 @@ Line wrap the file at 100 chars.                                              Th
 - Rename "Advanced settings" to "VPN Settings".
 - Move the "Split tunneling" menu item up a level from "VPN settings" to "Settings".
 - Remove purchasing information from release builds. This does not affect F-Droid builds.
+- Migrate Select Location view to compose.
 
 ### Fixed
 #### Android